HTTP/Server: Don't require `OnRespond' event to return a boolean value.
HTTP_Server_Process()'s sole purpose is to return true whilst more
input is needed before it's able to call the responder (which
hopefully makes up a response and sends it back to the client).

Currently it's also abused for stating whether the connection is
supposed to be kept open or not (as with HTTP/1.1). Hence, the
return value may have multiple meanings which depends on the case.