Server: Add an event triggered when the client requests data (EPOLLOUT).
EPOLLOUT is commonly used in edge-triggered mode. Refer to epoll(7).

This patch moves exception handling for connections from the
`ClientListener' to the `Server' module where it's used for EPOLLIN
as well as EPOLLOUT events.

The `onClientData' event is now called `onPull'.  Subscribing them
isn't anymore compulsory. They are getting only subscribed when
there is also a corresponding handler declared.