String: Rename `heap' to `mutable'.
While being valid for earlier revisions, the recent changes
resulted in a shift in its meaning.

Currently a string with `heap' set to false can mean that it is
either stack-allocated or refers to a string which is allocated on
the heap. In the latter case, `heap' only indicates that the
structure is not writable.

This technique allows greater efficiency because it reduces
unnecessary string duplication. It is employed in String_Slice(),
String_Trim(), Path_GetFilename(), Path_GetDirectory() amongst
others.