Revert "HTTP/Server: Detect HTTP/1.0 keep-alive connections."
This reverts commit 7e762ca1ecc3d54c3f73276b1416ff92683e320d.

When LightTPD is used as a proxy, it does not support keep-alive
connections. Therefore, it drops the "Connection" header but fails
to do the same for "Keep-Alive". This header is actually sent by the
browser and is just passed through.

It was falsely assumed that the presence of the "Keep-Alive" header
suffices, to qualify the client for supporting persistent
connections which is wrong as proven by this bug.