HTML/Tokenizer: Rewrite code.
The new variant should be a lot more memory-efficient. It is also
easier to maintain. The rewrite was necessary to make it possible
to implement DOCTYPE and CDATA in a clean way.

A disadvantage is that characters cannot be "fed" anymore and the
whole HTML file must be available in a string. Stream support was
dropped.